Add EPL
EPL (Emacs Package Library) is an abstraction layer on top of
package.el, to provide compatibility with its various API versions, and
to overcome many quirks and shortcomings.

The current EPL implementation deals only with the most recent
package.el API (aka defstruct package.el), but it will be backported to
older versions.