Rephrasing introduction to Cask
I found the Cask explanation confusing initially, since I have never
used Bundler. Since it's a dependency management tool, I thought it
would be easier if that was stated first, before making analogies.

I appreciate this is a minor cosmetic issue, but it was a hurdle for me
when I first started playing with Cask.