Support for --emacs command line option.
Previously, the emacs executable to be used had to be passed in as an
environment variable, or using heuristics within cask. While this works
under many circumstances, it can be a little erratic when running within
other environments (notably Emacs which can reset the EMACS environment
variable). This option allows the user to be extremely explicit.