Update Travis config
This change removes running the tests in Emacs 24.1 and 24.2.  I never
got the Travis config working with those two.  Probably because
emacs-web-server is only supported in >= 24.3.

Cask still supports those versions but they will not run on Travis.
It's better to have working tests on Travis in the most recent versions
of Emacs, than not having a working Travis at all.