Use HTTPS to connect to sources that support it
HTTP sends plaintext messages, so it leaks information and is
vulnerable to MITM attacks. ELPA, MELPA, and Marmalade all support
HTTPS. We should make use of it.

This changes the default addresses for those sources to use HTTPS. Users
that have built Emacs to handle TLS will be able to make use of these
sources, and Emacs will fall back to using regular HTTP if it isn't
configured to handle TLS.

This also updates the documentation to point to the HTTPS address of
those repositories.

This should close https://github.com/cask/cask/issues/323.