Improve handling on EMACS env vars.
Previously, the EMACS environment variable is ignored if Cask is running
inside Emacs, which was necessitated by Emacs-24 which changes alters
the value of this variable unhelpfully. As Emacs-25 does not change this
value, in this change we limit this behaviour to Emacs-24.

Additionally, we add support for a CASK_EMACS environment variable which
will work on both Emacs-24 and Emacs-25, although we document that the
on Emacs-25 the EMACS variable is prefered.