Don't test for "No" in the output.
In older versions of emacs (<=24), the error message is "no such file or
directory" but this got changed to "No such file or directory" in
emacs 25.

This just removes the "no" entirely from the error message, since it
really isn't important how it's spelled. Mainly, we just need to make
sure that the error for the file not being found was reported