`cask-cli--silent` handling illogic
The original intent of

(when (and (boundp silent) (not silent))
  (print))

is unclear as presumably if `silent` is not bound, we should print.

The upshot is any context in which cask-cli is not loaded,
e.g., any `cask exec` invocation, does not honor `cask-print`.

This is almost certainly incorrect, but it's too late to change.
This commit does not attempt to rectify, but rather clarify the logic.