Do we really need to test all versions?
The fact that we've allowed the CI to red-X for the last hundred runs
would suggest testing all 13 minor verisons is a false diligence.

The conao3/setup-cask step seems anachronistic.  We want to run bin/cask
from the checked out repo (that is what we want to test), not run
cask from a fetched copy of master.