Add implementation for package-build--get-timestamp
Upstream package-build added this new method which results in an error
when running cask since it's not implemented for package-directory-recipe.

The timestamp is used for reproducible builds based on commit
timestamp it seems from the upstream code. This is probably not very
useful for cask so just provide the current time here.