Fix Makefile compatibility with recent GNU Make
GNU Make will directly invoke simple commands[1], which causes the
following error when installing Cask:

    make: which: No such file or directory
    bin/cask install
    make: -c: No such file or directory
    make: *** [Makefile:47: /home/wilfred/src/cask/.cask/29.1/elpa] Error 127

This is because `which` is a shell built-in, and there's no guarantee
that there's a `which` binary available (e.g. my Arch Linux system
doesn't have one).

Use `command -v`, which is more portable, and use shell redirection so
Make never attempts to directly invoke it.

[1] https://stackoverflow.com/questions/17547625/how-to-use-shell-builtin-function-from-a-makefile/17550243#17550243